decode memo on activity page
